+++ example/shared-types-re/test/Spec.hs
@@ -1,0 +1,9 @@
+import OCaml.Export
+import Shared.Types.Reason.Types
+import Test.Hspec
+
+main :: IO ()
+main = do
+  mkGoldenFiles (Proxy :: Proxy SharedTypesPackage) 100 "test/golden"
+
+  hspec $ runGoldenSpec (Proxy :: Proxy SharedTypesPackage) 100 "test/golden"