+++ test/interface/golden/__tests__/product/Key_spec.ml
@@ -1,0 +1,12 @@
+let () =
+  AesonSpec.goldenDirSpec
+    Key.decodeSqlKey
+    Key.encodeSqlKey
+    "sqlKey"
+    "golden/product/SqlKey";
+
+  AesonSpec.goldenDirSpec
+    Key.decodeUserId
+    Key.encodeUserId
+    "userId"
+    "golden/product/UserId";